+++ src/nostr/DefaultRelays.ts
@@ -4,7 +4,7 @@ export const defaultRelays: RelaySetting[]  = [
     {relayUrl: "wss://nostr.wine", read: false, write: false},
     {relayUrl: "wss://filter.nostr.wine", read: true, write: true},
     {relayUrl: "wss://eden.nostr.land", read: true, write: false},
-    {relayUrl: "wss://relay.nostr.band/", read: true, write: false},
+    {relayUrl: "wss://relay.nostr.band/", read: false, write: false},
     {relayUrl: "wss://nostr-pub.wellorder.net", read: false, write: true},
     {relayUrl: "wss://relay.nostr.band/all", read: false, write: true},
     {relayUrl: "wss://relay.snort.social", read: false, write: true},
